pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incident_response"
version = "0.1.0"
description = "Mini Incident Response Copilot Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "pydantic>=2.0.0",
    "openai>=1.0.0",  # For inference script
    "numpy>=1.24.0",  # For metrics calculations
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m incident_response.server.app
server = "incident_response.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["incident_response", "incident_response.server"]
package-dir = { "incident_response" = ".", "incident_response.server" = "server" }